food_matcher/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
numpy==1.26.4
pymysql==1.1.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
tqdm==4.66.4
